Renamed the ILogParser class family to INightlyProcessLog
