--- HousingFinanceInterimApi/V1/Gateways/Interface/ILogParserGateway.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Amazon.CloudWatchLogs.Model;
-
-namespace HousingFinanceInterimApi.V1.Gateway.Interfaces
-{
-    public interface ILogParserGateway
-    {
-        Task UpdateDatabaseWithResults(string logGroupName, List<List<ResultField>> queryResults);
-    }
-}
