+++ HousingFinanceInterimApi/V1/UseCase/Interfaces/INightlyProcessLogUseCase.cs
@@ -1,12 +1,15 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using System;
+using HousingFinanceInterimApi.V1.Boundary.Response;
 using HousingFinanceInterimApi.V1.Domain;
+using System.Collections.Generic;
+using System;
+using System.Threading.Tasks;
 
 namespace HousingFinanceInterimApi.V1.UseCase.Interfaces
 {
     public interface INightlyProcessLogUseCase
     {
-        Task<List<NightlyProcessLogResponse>> ExecuteAsync(DateTime createdDate);
+        Task<StepResponse> ExecuteAsync();
+
+        Task<IList<NightlyProcessLogResponse>> ExecuteAsync(DateTime createdDate);
     }
 }
